fix: retail prices are in dollars, not cents — remove /100 conversion
Catalog prices are in dollars (20, 12). The original 100x inflation
was from the frontend Math.round(price*100) which was already fixed.
The backend /100 was overcorrecting.

+++ services/api/src/routes/checkin/retailLedger.ts
@@ -11,7 +11,7 @@ const AddRetailItemsSchema = z.object({
     sku: z.string().optional().nullable(),
     name: z.string().min(1),
     quantity: z.number().int().positive(),
-    unitPrice: z.number().int().nonnegative(),
+    unitPrice: z.number().nonnegative(),
   })).min(1),
 });
